1561 - refactor
